Clean ups.
	Compiling with X11.
	* configure.ac: Use URBI_PROG_CXX.
	Use AC_ENABLE_SHARED.
	Use AC_PATH_XTRA.
	* src/examples/Makefile.am: Use the results.

	* src/utils/urbicycle.cpp: Formatting changes.
	Remove unused variables.
	Add return values where missing.
	Add cases to incomplete switches.
	delete 0 is valid C++.
	Prefer the anonymous namespace to static.
	Initialize members in the same order as the one of the declaration.
	Prefer stack resident variables to heap values.
	* src/liburbi/usyncclient.cpp: .
	* src/liburbi/lockable.h: .
	* src/liburbi/uobject/uvar.cc: .
	* src/liburbi/uobject/uobject.template.h: .
	* src/liburbi/uobject/common_uvalue.cc: .
	* src/examples/urbisendsound.cpp: .
	* src/examples/urbiplay.cpp: .
	* src/examples/urbitalkie.cpp: .
	* src/examples/urbirecord.cpp: .
	* src/examples/urbiballtrackinghead.cpp: .
	* src/examples/monitor.h: .


git-svn-id: file:///srv/svn/liburbi-cpp/trunk@179 b8db33d8-8104-0410-b827-8403f563b8eb

+++ sdk-remote/src/examples/urbisendsound.cpp
@@ -67,7 +67,7 @@ main(int argc, char * argv [])
       convert(s, snd);
 
       uc.setCallback(endProgram,"end");
-      printf("sending %d bytes\n",st.st_size);
+      printf("sending %d bytes\n", static_cast<int>(st.st_size));
       uc.sendSound("speaker", snd,"end");
       printf("done, waiting for end of play notification\n");
     }
